feat: 쿼리 및 뮤테이션 분리
