+++ src/entities/post/api/queries.ts
@@ -1,8 +1,7 @@
 import { queryOptions } from "@tanstack/react-query"
 
-import { CreatePostDto, FetchPostsBySearchParams, FetchPostsByTagParams, FetchPostsParams, Post } from "../model/types"
+import { FetchPostsBySearchParams, FetchPostsByTagParams, FetchPostsParams } from "../model/types"
 import { postApi } from "."
-import { queryClient } from "../../../shared/api/query-client"
 
 export const postQueries = {
   all: () => ["posts"] as const,
@@ -43,36 +42,4 @@ export const postQueries = {
       queryKey: [...postQueries.all(), "tag"],
       queryFn: postApi.fetchTags,
     }),
-}
-
-export const postMutations = {
-  addMutation: () => ({
-    mutationKey: [...postQueries.all(), "add"] as const,
-    mutationFn: (post: CreatePostDto) => postApi.addPost(post),
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: postQueries.all(),
-      })
-    },
-  }),
-
-  updateMutation: () => ({
-    mutationKey: [...postQueries.all(), "update"] as const,
-    mutationFn: ({ id, post }: { id: number; post: Partial<Post> }) => postApi.updatePost(id, post),
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: postQueries.all(),
-      })
-    },
-  }),
-
-  deleteMutation: () => ({
-    mutationKey: [...postQueries.all(), "delete"] as const,
-    mutationFn: (id: number) => postApi.deletePost(id),
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: postQueries.all(),
-      })
-    },
-  }),
 }
